+++ test/query.test.js
@@ -4001,4 +4001,14 @@ describe('Query', function() {
     await Test.findOneAndUpdate({}, { name: 'bar' });
     assert.ok(!('projection' in lastOptions));
   });
+  it('should provide a clearer error message when sorting with empty string', async function() {
+    const testSchema = new Schema({
+      name: { type: String }
+    });
+
+    const Error = db.model('error', testSchema);
+    await assert.rejects(async() => {
+      await Error.find().sort('-');
+    }, { message: 'Invalid field "" passed to sort()' });
+  });
 });